Revert "Increase data width of poc for 4K sequences"
This reverts commit 97c4e942da9520b3a8542fe75bcabfb7fd4844d9.

+++ libavcodec/hevc.h
@@ -1018,7 +1018,7 @@ typedef struct Mv {
 typedef struct MvField {
     Mv mv[2];
 #ifdef TEST_MV_POC
-    int32_t poc[2];
+    int16_t poc[2];
     uint32_t pred_flag;
 #else
     uint8_t pred_flag;
